+++ silx/gui/plot/_utils/test/__init__.py
