+++ silx/gui/plot/_utils/ticklayout.py
